Only call _parse_version() for lines that contain VERSION
I.e. move the check for 'VERSION' from inside _parse_version() to the loop
of check_install(). This saves a function call for each source line that
does not contain 'VERSION', yielding a whopping 20% speedup for
check_install().

Note that _parse_version() still returns undef when the code chunk does
not contain 'VERSION', because the main regexp won't match.