Port $VERSION parsing in check_install() to Module::Metadata
This replaces the hand-made $VERSION parsing with its canonical toolchain
implementation, and fixes numerous issues with how $VERSION was parsed.
Module::Metadata 1.000_005 is required for its new_from_handle method.

In particular it is now aware of the surrounding package, and tests have
been added for files without and with several package statements.

As a result of this change, the t/02_Parse_Version.t test file is now
obsolete and is removed from the distribution.